requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.3
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
filelock==3.12.0
flatbuffers==23.3.3
fonttools==4.39.3
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
lit==16.0.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnx==1.13.1
onnxruntime==1.14.1
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==3.20.3
psutil==5.9.5
pycocotools==2.0.6
pycodestyle==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.29.0
scipy==1.10.1
seaborn==0.12.2
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@567662b0fd33ca4b022d94d3b8de896628cd32dd
sentry-sdk==1.21.1
six==1.16.0
sympy==1.11.1
thop==0.1.1.post2209072238
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.89
urllib3==1.26.15
